Collision works!
It needs to be tested more but currently if you click on the ship, shields go up. (Just something simple to test).
It uses RectCollision to test for collision. It is current implemented via HelmScreen ONLY. I'll merge it into the ship
class shortly.


Problems: For whatever reason, I could not get the collision box to rotate around the origin (27,11) we were using
for the ship to rotate around. So everything rotates now around 0,0 (relative to the ship). So it all looks kinda
funky. But we can live with this...for now. As another result of this change, the beam code doesn't work properly.
I just don't care right now.

I changed the enterprise90t sprite to fit better, there is no excess space. I adjusted the shield texture as well.

As always, I cleaned up some code as I came across it and removed defunct/unused code for the regular housekeeping.