BIG CHANGES!!! :)
A basic version of scenarios has  been implemented, I have a few small changes to implement before we start to use them,
but the game world is now a Scenario. The Game1 class has been heavily reduced. I need to tweak a few things but for now
it works. DefaultScenario is now a class which contains the majority of Game1. Everything has been updated to work with
the new class layout. Game1 cannot switch scenarios yet but the abstraction is almost there. Its a few lines I'll change
after this, but "commit early, commit often".

This change helps lead us torwards modularizing the code and enables us to begin coding a level. A Kobayashi Maru scenario
is not far off at all.